Move import functions into separate file
